fix: TEMPORARY. ssr ruins voteStatus return
For some stupid reason, when ssr is turned on for index our posts
resolver cannot use req.session.userId. We use req.session.userId in
order to set the voteStatus in our SQL query, which is in turn used to
set the coloring and functionality of the upper and downer buttons. So,
until we can figure out why ssr ruins the cookie for our posts resolver,
index is not ssr